requirements.txt
feedparser==6.0.11
requests==2.31.0
google-generativeai==0.3.2
